Le fichier readme

+++ smartlogi/src/main/java/com/smartlogi/services/LivreurService.java
@@ -2,6 +2,8 @@ package com.smartlogi.services;
 
 import com.smartlogi.entities.Livreur;
 import com.smartlogi.repositories.LivreurRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.Optional;
 
@@ -9,17 +11,24 @@ public class LivreurService {
 
     private LivreurRepository livreurRepository;
 
-    // Injection par setter
-    public void setLivreurRepository(LivreurRepository livreurRepository) {
+
+    public void setLivreurRepository (LivreurRepository livreurRepository){
         this.livreurRepository = livreurRepository;
+
     }
+
+
+
 
     public Livreur ajouterLivreur(Livreur livreur) {
-        return livreurRepository.save(livreur);
+         return livreurRepository.save(livreur);
+
     }
 
-    public List<Livreur> tousLesLivreurs() {
-        return livreurRepository.findAll();
+    public void tousLesLivreurs() {
+        List<Livreur> livreurs = livreurRepository.findAll();
+        livreurs.forEach(System.out::println);
+
     }
 
     public Optional<Livreur> trouverParId(Long id) {
@@ -27,7 +36,12 @@ public class LivreurService {
     }
 
     public void supprimerLivreur(Long id) {
-        livreurRepository.deleteById(id);
+       Optional <Livreur> find = livreurRepository.findById(id);
+        if(find.isPresent()){
+            livreurRepository.deleteById(id);
+        }
+        else
+            System.out.print("id n existe pas");
     }
 
     public Livreur mettreAJourLivreur(Livreur livreur) {
